toolkit/mozapps/update/tests/chrome/test_0091_installed.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: installed"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

const TESTS = [ {
  pageid: PAGEID_INSTALLED,
  buttonClick: "finish"
} ];

function runTest() {
  debugDump("entering");

  Services.prefs.setBoolPref(PREF_APP_UPDATE_SHOW_INSTALLED_UI, true);
  let patches = getLocalPatchString("complete", null, null, null, null, null,
                                    STATE_PENDING);
  let updates = getLocalUpdateString(patches, null, null, null,
                                     Services.appinfo.version,
                                     Services.appinfo.platformVersion);
  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);
  writeUpdatesToXMLFile(getLocalUpdatesXMLString(""), false);
  writeStatusFile(STATE_SUCCEEDED);

  testPostUpdateProcessing();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
